+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/info/License.kt
@@ -1,0 +1,8 @@
+package io.bkbn.kompendium.oas.info
+
+import java.net.URI
+
+data class License(
+  var name: String,
+  var url: URI? = null
+)