+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/AnyOfSchema.kt
@@ -1,0 +1,3 @@
+package io.bkbn.kompendium.oas.schema
+
+data class AnyOfSchema(val anyOf: List<ComponentSchema>) : ComponentSchema