+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/EnumSchema.kt
@@ -1,0 +1,8 @@
+package io.bkbn.kompendium.oas.schema
+
+data class EnumSchema(
+  val `enum`: Set<String>,
+  override val default: Any? = null
+) : TypedSchema {
+  override val type: String = "string"
+}