+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/ObjectSchema.kt
@@ -1,0 +1,8 @@
+package io.bkbn.kompendium.oas.schema
+
+data class ObjectSchema(
+  val properties: Map<String, ComponentSchema>,
+  override val default: Any? = null
+) : TypedSchema {
+  override val type = "object"
+}