+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/SimpleSchema.kt
@@ -1,0 +1,6 @@
+package io.bkbn.kompendium.oas.schema
+
+data class SimpleSchema(
+  override val type: String,
+  override val default: Any? = null
+) : TypedSchema