+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/TypedSchema.kt
@@ -1,0 +1,6 @@
+package io.bkbn.kompendium.oas.schema
+
+sealed interface TypedSchema : ComponentSchema {
+  val type: String
+  override val default: Any?
+}